feat(ui): implement modal authentication with login and register functionality refactor(ui): optimize layout & pages components
