+++ frontend/src/app/(main)/layout.tsx
@@ -1,4 +1,4 @@
-import Footer from '@components/Footer/Footer'
+import Footer from '@components/Footer'
 import Header from '@components/Header'
 import type { Metadata } from 'next'
 
@@ -7,9 +7,9 @@ export const metadata: Metadata = {
 	description: 'An app to share and post texts',
 }
 
-export default function MainLayout({ children }: { children: React.ReactNode }) {
+export default function MainLayout({ children }: React.PropsWithChildren) {
 	return (
-		<div className='max-w-[1270px] m-[0_auto] font-urbanist grid grid-rows-[10vh_80vh]'>
+		<div className='max-w-[1270px] m-[0_auto] grid grid-rows-[10vh_minmax(80vh,auto)]'>
 			<Header />
 			{children}
 			<Footer />
