+++ frontend/src/app/layout.tsx
@@ -1,12 +1,12 @@
+import ModalRouter from '@components/ModalRouter'
 import './globals.css'
 
-export default function Layout({ children }: { children: React.ReactNode }) {
+export default async function Layout({ children }: React.PropsWithChildren) {
 	return (
 		<html lang='en'>
-			<body>
-				<div className='bg-gray-100 dark:bg-gray-950 min-h-screen text-gray-800 dark:text-gray-100'>
-					{children}
-				</div>
+			<body className='bg-gray-100 dark:bg-gray-950 min-h-screen text-gray-800 dark:text-gray-100 font-urbanist'>
+				{children}
+				<ModalRouter />
 			</body>
 		</html>
 	)
