+++ frontend/src/app/storage/layout.tsx
@@ -1,4 +1,4 @@
-import Footer from '@components/Footer/Footer'
+import Footer from '@components/Footer'
 import Header from '@components/Header'
 import type { Metadata } from 'next'
 
@@ -7,9 +7,9 @@ export const metadata: Metadata = {
 	description: 'Storage of your texts',
 }
 
-export default function StorageLayout({ children }: { children: React.ReactNode }) {
+export default function StorageLayout({ children }: React.PropsWithChildren) {
 	return (
-		<div className='max-w-[1270px] m-[0_auto] font-urbanist grid grid-rows-[10vh_auto]'>
+		<div className='max-w-[1270px] m-[0_auto] grid grid-rows-[10vh_minmax(80vh,auto)]'>
 			<Header />
 			{children}
 			<Footer />
