+++ frontend/src/app/text/layout.tsx
@@ -1,4 +1,4 @@
-import Footer from '@components/Footer/Footer'
+import Footer from '@components/Footer'
 import Header from '@components/Header'
 import type { Metadata } from 'next'
 
@@ -9,7 +9,7 @@ export const metadata: Metadata = {
 
 export default function TextLayout({ children }: { children: React.ReactNode }) {
 	return (
-		<div className='max-w-[1270px] m-[0_auto] font-urbanist grid grid-rows-[10vh_80vh]'>
+		<div className='max-w-[1270px] m-[0_auto] grid grid-rows-[10vh_minmax(80vh,auto)]'>
 			<Header />
 			{children}
 			<Footer />
