+++ frontend/src/components/Header/LogOutBtn.tsx
