+++ frontend/src/components/ModalRouter/Modals/Login.tsx
@@ -4,24 +4,23 @@ import Button from '@components/shared/Button'
 import Input from '@components/shared/Input'
 import { usePathname, useRouter } from 'next/navigation'
 import { useState } from 'react'
-import Modal from './Modal'
+import ModalWrapper from '../ModalWrapper'
 
-export default function Register() {
+export default function Login() {
 	const [username, setUsername] = useState('')
-	const [email, setEmail] = useState('')
 	const [password, setPassword] = useState('')
 
 	const router = useRouter()
 	const pathname = usePathname()
 
 	return (
-		<Modal
+		<ModalWrapper
 			onClose={() => {
 				router.push(pathname)
 			}}
 		>
-			<div className='flex justify-center items-center bg-gray-100 dark:bg-gray-950 p-10 rounded-md'>
-				<div className='grid grid-rows-4 gap-y-5'>
+			<main className='flex justify-center items-center bg-gray-100 dark:bg-gray-950 p-10 rounded-md'>
+				<div className='grid grid-rows-3 gap-y-5'>
 					<div className='flex flex-col'>
 						<span className='dark:text-gray-200 text-gray-700 text-md'>Username</span>
 						<Input
@@ -30,16 +29,6 @@ export default function Register() {
 							className='w-64'
 							ariaLabel='Username input field'
 							onChange={e => setUsername(e.target.value)}
-						/>
-					</div>
-					<div className='flex flex-col'>
-						<span className='dark:text-gray-200 text-gray-700 text-md'>Email</span>
-						<Input
-							name={'email'}
-							value={email}
-							className='w-64'
-							ariaLabel='Email input field'
-							onChange={e => setEmail(e.target.value)}
 						/>
 					</div>
 					<div className='flex flex-col'>
@@ -54,13 +43,13 @@ export default function Register() {
 						/>
 					</div>
 					<Button
-						name={'Register'}
+						name={'Login'}
 						className='self-end bg-blue-400 dark:bg-blue-500 hover:bg-blue-500 dark:hover:bg-blue-600'
-						ariaLabel='Register button'
-						onClick={() => console.log('Register')}
+						ariaLabel='Login button'
+						onClick={() => console.log('Login')}
 					/>
 				</div>
-			</div>
-		</Modal>
+			</main>
+		</ModalWrapper>
 	)
 }
