--- frontend/src/components/Modals/index.tsx
@@ -1,2 +1,0 @@
-export { default as Login } from './Login'
-export { default as Register } from './Register'
